+++ lib/rbs/merge/node_type_normalizer.rb
@@ -288,4 +288,3 @@ module Rbs
     end
   end
 end
-