+++ spec/integration/reproducible_merge_spec.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require "spec_helper"
 require "rbs/merge"
 require "ast/merge/rspec/shared_examples"