+++ spec/spec_helper.rb
@@ -1,36 +1,4 @@
 # frozen_string_literal: true
 
-# External RSpec & related config
-require "kettle/test/rspec"
-require "ast/merge/rspec"
-
-# Internal ENV config
-require_relative "config/debug"
-
-# Config for development dependencies of this library
-# i.e., not configured by this library
-#
-# Simplecov & related config (must run BEFORE any other requires)
-# NOTE: Gemfiles for older rubies won't have kettle-soup-cover.
-#       The rescue LoadError handles that scenario.
-begin
-  require "kettle-soup-cover"
-  require "simplecov" if Kettle::Soup::Cover::DO_COV # `.simplecov` is run here!
-rescue LoadError => error
-  # check the error message and re-raise when unexpected
-  raise error unless error.message.include?("kettle")
-end
-
-# this library
-require "rbs/merge"
-
-# Support files (including dependency tags for conditional test execution)
-Dir[File.join(__dir__, "support", "**", "*.rb")].each { |f| require f }
-
-RSpec.configure do |config|
-  config.before do
-    # Speed up polling loops
-    allow(described_class).to receive(:sleep) unless described_class.nil?
-  end
-end
-
+# Much of the config is loaded automatically by the .rspec config via spec_thin_helper
+# This file is for gem-specific requires that individual specs may need